+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ChecklistDao.java
@@ -1,6 +1,7 @@
 package br.com.zalf.prolog.webservice.frota.checklist;
 
 import br.com.zalf.prolog.webservice.frota.checklist.OLD.Checklist;
+import br.com.zalf.prolog.webservice.frota.checklist.model.ChecklistListagem;
 import br.com.zalf.prolog.webservice.frota.checklist.model.FiltroRegionalUnidadeChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.farol.DeprecatedFarolChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.insercao.ChecklistInsercao;
@@ -94,34 +95,33 @@ public interface ChecklistDao {
     Checklist getByCod(@NotNull final Long codChecklist) throws SQLException;
 
     /**
-     * Busca todos os checklists, respeitando os filtros aplicados (recebidos por parâmetro).
+     * Busca os checklists realizados por um colaborador.
      *
-     * @return uma {@link List<Checklist> lista de checklists}.
+     * @return uma {@link List<ChecklistListagem> lista de checklists}.
      * @throws SQLException caso não seja possível realizar a busca.
      */
     @NotNull
-    List<Checklist> getAll(@NotNull final Long codUnidade,
-                           @Nullable final Long codEquipe,
-                           @Nullable final Long codTipoVeiculo,
-                           @Nullable final String placaVeiculo,
-                           final long dataInicial,
-                           final long dataFinal,
-                           final int limit,
-                           final long offset,
-                           final boolean resumido) throws SQLException;
+    List<ChecklistListagem> getListagemByColaborador(@NotNull final Long codColaborador,
+                                                     @NotNull final LocalDate dataInicial,
+                                                     @NotNull final LocalDate dataFinal,
+                                                     final int limit,
+                                                     final long offset) throws Throwable;
 
     /**
-     * Busca os checklists realizados por um colaborador.
+     * Busca todos os checklists, respeitando os filtros aplicados (recebidos por parâmetro).
      *
-     * @return uma {@link List<Checklist> lista de checklists}.
+     * @return uma {@link List<ChecklistListagem> lista de checklists}.
      * @throws SQLException caso não seja possível realizar a busca.
      */
-    List<Checklist> getByColaborador(@NotNull final Long cpf,
-                                     @NotNull final Long dataInicial,
-                                     @NotNull final Long dataFinal,
-                                     final int limit,
-                                     final long offset,
-                                     final boolean resumido) throws SQLException;
+    @NotNull
+    List<ChecklistListagem> getListagem(@NotNull final Long codUnidade,
+                                        @Nullable final Long codEquipe,
+                                        @Nullable final Long codTipoVeiculo,
+                                        @Nullable final Long codVeiculo,
+                                        @NotNull final LocalDate dataInicial,
+                                        @NotNull final LocalDate dataFinal,
+                                        final int limit,
+                                        final long offset) throws Throwable;
 
     /**
      * Busca as regionais e unidades que o colaborador de código fornecido tem acesso. Isso é verificado com base na
@@ -164,4 +164,40 @@ public interface ChecklistDao {
      * @throws Throwable Caso ocorrer algum erro na busca dos dados.
      */
     boolean getChecklistDiferentesUnidadesAtivoEmpresa(@NotNull final Long codEmpresa) throws Throwable;
+
+    /**
+     * Busca todos os checklists, respeitando os filtros aplicados (recebidos por parâmetro).
+     *
+     * @return uma {@link List<Checklist> lista de checklists}.
+     * @throws SQLException caso não seja possível realizar a busca.
+     * @deprecated at 2020-06-08.
+     * Use {@link ChecklistDao#getListagem(Long, Long, Long, String, LocalDate, LocalDate, int, long)} instead.
+     */
+    @NotNull
+    @Deprecated
+    List<Checklist> getAll(@NotNull final Long codUnidade,
+                           @Nullable final Long codEquipe,
+                           @Nullable final Long codTipoVeiculo,
+                           @Nullable final String placaVeiculo,
+                           final long dataInicial,
+                           final long dataFinal,
+                           final int limit,
+                           final long offset,
+                           final boolean resumido) throws SQLException;
+
+    /**
+     * Busca os checklists realizados por um colaborador.
+     *
+     * @return uma {@link List<Checklist> lista de checklists}.
+     * @throws SQLException caso não seja possível realizar a busca.
+     * @deprecated at 2020-06-08.
+     * Use {@link ChecklistDao#getListagemByColaborador(Long, LocalDate, LocalDate, int, long)} instead.
+     */
+    @Deprecated
+    List<Checklist> getByColaborador(@NotNull final Long cpf,
+                                     @NotNull final Long dataInicial,
+                                     @NotNull final Long dataFinal,
+                                     final int limit,
+                                     final long offset,
+                                     final boolean resumido) throws SQLException;
 }