+++ src/main/java/br/com/zalf/prolog/webservice/integracao/router/Router.java
@@ -6,6 +6,7 @@ import br.com.zalf.prolog.webservice.customfields.CampoPersonalizadoDao;
 import br.com.zalf.prolog.webservice.customfields._model.CampoPersonalizadoParaRealizacao;
 import br.com.zalf.prolog.webservice.frota.checklist.ChecklistResource;
 import br.com.zalf.prolog.webservice.frota.checklist.OLD.Checklist;
+import br.com.zalf.prolog.webservice.frota.checklist.model.ChecklistListagem;
 import br.com.zalf.prolog.webservice.frota.checklist.model.TipoChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.farol.DeprecatedFarolChecklist;
 import br.com.zalf.prolog.webservice.frota.checklist.model.insercao.ChecklistInsercao;
@@ -343,6 +344,54 @@ public abstract class Router implements OperacoesIntegradas {
             return integradorProLog
                     .getTodosChecklists(codUnidade, codEquipe, codTipoVeiculo, placaVeiculo, dataInicial, dataFinal,
                             limit, offset, resumido);
+        }
+    }
+
+    @NotNull
+    @Override
+    public List<ChecklistListagem> getListagemByColaborador(@NotNull final Long codColaborador,
+                                                            @NotNull final LocalDate dataInicial,
+                                                            @NotNull final LocalDate dataFinal,
+                                                            final int limit,
+                                                            final long offset) throws Throwable {
+        if (getSistema() != null) {
+            return getSistema().getListagemByColaborador(codColaborador, dataInicial, dataFinal, limit, offset);
+        } else {
+            return integradorProLog.getListagemByColaborador(codColaborador, dataInicial, dataFinal, limit, offset);
+        }
+    }
+
+    @NotNull
+    @Override
+    public List<ChecklistListagem> getListagem(@NotNull final Long codUnidade,
+                                               @Nullable final Long codEquipe,
+                                               @Nullable final Long codTipoVeiculo,
+                                               @Nullable final Long codVeiculo,
+                                               @NotNull final LocalDate dataInicial,
+                                               @NotNull final LocalDate dataFinal,
+                                               final int limit,
+                                               final long offset) throws Throwable {
+        if (getSistema() != null) {
+            return getSistema()
+                    .getListagem(
+                            codUnidade,
+                            codEquipe,
+                            codTipoVeiculo,
+                            codVeiculo,
+                            dataInicial,
+                            dataFinal,
+                            limit,
+                            offset);
+        } else {
+            return integradorProLog
+                    .getListagem(codUnidade,
+                            codEquipe,
+                            codTipoVeiculo,
+                            codVeiculo,
+                            dataInicial,
+                            dataFinal,
+                            limit,
+                            offset);
         }
     }